requirements.txt
Django==4.0.3
django-crispy-forms==1.14.0
entrypoints==0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
happytransformer==3.0.0
names_generator
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
praat-parselmouth==0.4.3
Pillow==8.4.0
protobuf==3.19.4
python-pptx
scikit-image==0.18.3
scikit-learn
scipy
seaborn 
SQLAlchemy
SpeechRecognition==3.10.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0

json-tricks
librosa
pydub
noisereduce
PyAudio
fer